Add server
This adds a basic server that runs on port 3000.
We could use a library or framework for this but we want to keep this
project as easy to maintain as possible.
Plus, this isn't a particularly complex app and to begin with it will
have one route so I think the out of the box Node Server will suffice.