Use `@octokit/app` as the main Octokit dependency
GitHub encourages the use of the GitHub App workflow when a system needs
to interact with the GitHub API on its own behalf, rather than using
personal access tokens which need to be regenerated on a regular basis.

This library exposes the GitHub API in a way that's indended to be used
for this workflow.

See
https://docs.github.com/en/apps/creating-github-apps/authenticating-with-a-github-app/about-authentication-with-a-github-app
for more information about authentication for GitHub apps, including
when an application should authenticate as itself rather than as a
user.