Use `OctokitApp` to display repository list
This setup iterates over all installations of the GitHub app and
displays their repositories in the previously made table.

Iterating over multiple installations rather than specifying it
explicitly will allow us to support repositories we manage in other
organisations by getting them to install Towtruck as a GitHub app in
their organisation settings and granting access to the repositories that
we maintain for them.

However, currently we only want to support single-account installations.
There doesn't seem to be a neat way to get the installation ID from an
account name, so we will use `eachInstallation` to loop (hopefully once)
and just take the first (hopefully only) element from `installations` so
that we can have more meaningful template names in Nunjucks.

We can enforce this one-installation approach through GitHub by
configuring the app to be "Only on this account" when registering the app.