Use Tailwind for basic styling
[Tailwind](https://tailwindcss.com/) is a CSS framework designed to
provide a collection of composable classes that can be combined to style
elements in a visually coherent way.

Typically, Tailwind is used a `devDependency` and built to ensure a
minimal amount of CSS is used. However, in the spirit of being as
straightforward to maintain as possible, instead this just includes the
Tailwind CDN script in the `index.njk` template.

The visual changes include:
- Making the 'Towtruck' title prominent in a header bar
- A card-style interface showing the table of repositories for each GitHub installation
- A basic table layout with alternating row colours for easier visual
  identification.