Extract data from Renovate Dependency Dashboard issue
The `getDependenciesForRepo` function gets the list of issues for the
given repo and finds the first that's authored by the Renovate bot that
isn't a pull request.

Assuming that this is the Dependency Dashboard, it then iterates through
the body of the issue to parse dependencies.

Renovate currently formats the dashboard so that dependencies are
presented like so:

```md
- `some-dependency 1.2.3`
- `some-dependency-with-version-range ^2.3.4`
- `some-dependency-with-squiggly-version "~> v5"`
```

If there are no dependencies in the dashboard or this issue doesn't
exist, an empty array is returned.