Add script to fetch all repos from Github
This script pulls in all repos for the org where Towtruck is installed.
It pulls the relevant properties from the repo data and pushes this info
to an array. We ignore any archived repos which reduces the number of
repos by >50%.
We then write this repo data to a JSON file to be used later.
I couldn't work out a way to unit test the function that didn't result
in tests that were harder to read than the function itself. The mocking
in Node's built in test runner is still experimental so we may need to
introduce a standalone test runner (Vitest, Jest) to be able to mock
properly.