Update mapping function
This function won't be used to map repos from the Github API to the UI
anymore but from our stored repos to the UI repos.
This means we can just convert the date from ISO8601 to a JS Date
object. We retain the total number of repos as this may come in handy
later.