Fix fetching dependencies
I broke this with #27.
I switched from passing repos as a second parameter to passing it as
part of the object in the first parameter in the consuming function but
not in the function signature itself.
This format follows the signature the Octokit builtin functions use so I
think it makes sense to align our approach with theirs.