Add open PRs to dashboard
It will be useful to know how many open PRs a repo has. In future we can
make this even more useful but showing a column for 'Open Renovate PRs'.

In fetchAllRepos we are doing a `Promise.all` on the secondary requests
for data so the requests can run concurrently.
The seeding script takes a long time to run at the moment but with
change their shouldn't be too much of an extra penalty for further
concurrent API requests.