Move `Dependency` to `model` module
Dependency information can come from multiple sources - the list of
dependencies for a particular repository might come from Renovate, but
information about newer versions or the end-of-life date for a
particular version might come from elsewhere. Moving the `Dependency`
class makes it clearer that it's not intrinsically tied to Renovate.