Add data script to retrieve all end-of-life info for found dependencies
The data script goes through all the persisted repositories and collates
their dependencies. For each one, a query to the https://endoflife.date
API is performed, and if the API returns data, it gets stored.

The `npm run seed` command has been updated to perform two subcommands,
one for each data script.

This allows both the `seed` script to fully refresh the data, but also
for individual JSON files to be recreated, which is quicker.