Describe utils in with JS Doc
JS Doc is a widely used method of documenting Javascript.
It is widely supported by IDEs and runs in IDEs so doesn't require an
application depedencies - it's on the developer to keep their env up to
date.
It gives many of the benefits of Typescript without the learning curve
and the build step. It can (and should) be added optionally to improve
documentation and developer experience but it is optional won't effect
the run time code.