Add seeding
This adds a seeding job.
Using live data is hard to assert against and means we have to run the
seeding job which takes time and uses rate-limit tokens.
Instead we copy across some test data we can use. This means the data
will always be consistent and we can test against it more easily.
We can't do this in a setup job [as Playwright
recommends](https://playwright.dev/docs/test-global-setup-teardown)
because the app depends on the JSON existing before it can start.
Instead we have to copy and paste the seed data across before we start
the app.