Update dependency list to display information about new versions
When available, dependencies in the list will be colour-coded using
Tailwind colour names:
- `sky` for up-to-date dependencies
- `amber` for dependencies with minor version updates
- `orange` for dependencies with major version updates
- `red` for dependencies that are no longer supported
- `stone` for dependencies for which no information is available

The capsules containing the dependency information will use these
colours.

Additionally, for out-of-date dependencies, the latest version is
displayed in a separate, prominently coloured capsule, with an icon as a
secondary indicator of the outdatedness of the dependency. These icons
are class names for [Boxicons](https://boxicons.com/) icons:
- `bx-up-arrow-circle` for minor updates
- `bxs-up-arrow-circle` for major updates
- `bxs-x-circle` for end-of-life