Adjust mapping of repo
Previously we mapped the repo from its API format to the format we want
in our JSON just before pushing into into our array. This meant we add
to re-add properties that we just added to the repo again (eg
Dependencies).
Now we convert it from the API shape to our JSON shape then add the
custom properties. This means we can make changes in getOpenPRsForRepo
to add properties without having to update `mapRepoFromApiForStorage`
too.