Include `date-fns` dependency
This dependency introduces multiple utility functions for manipulating
and presenting dates which will be helpful for presenting things such as
the time left before support ends for a dependency or the age of an
issue.

This is the most popular dependency for date utilities and has no
downstream dependencies, and given that code that manipulates dates
and/or time are notoriously prone to gotchas, it seems like a good
candidate for inclusion despite our preference to minimise dependencies
for this project in general.