Make information in sortable columns sticky within their row
Each row in the table can become quite tall when the list of
dependencies is expanded. This allows information in the sortable
columns to not scroll off the top of the view until the row itself is
leaving the view, so that the information is visible most of the time.