Refactor seeding logic to be testable and reusable
The logic used to fetch and store information about repositories and
dependencies will need to be reused to allow webhooks to make changes to
the database. Using the logic in the application rather than in one-off
seeding scripts makes it more important that it is tested to avoid
unexpected changes to behaviour. This moves the actual seed scripts to
wrapper scripts in the `seed` folder, which call the methods in the
original files that have been refactored to be more testable.