Avoid the need for `nvm` to be installed in the GitHub Action workflows
`nvm` no longer seems to be present when running Node in a GitHub
Action. However, the correct Node version is already extracted in
previous steps, so it's not actually needed here. Therefore, the setup
script can just call `npm install` rather than delegate to
`script/bootstrap` in a CI environment, and will do this if the first
argument is `--ci`.