Configure GitHub OAuth setup
This defines the `REDIRECT_URL_BASE` environment variable which can be
used to configure which callback URL Towtruck provides when using the
OAuth authentication flow.

Octokit already provides as part of its middleware the ability to
authenticate as a GitHub user, so Towtruck just needs to be able to hook
into this process.