Introduce login page Nunjucks template
Towtruck should expose a login page rather than immediately redirecting
to the GitHub auth page, both to avoid surprising the user and also to
provide a landing area if more login methods are supported in the
future.

To make sure that this login page looks the same as the index page,
`index.njk` has been split:
- `base.njk` contains the `<head>` and outer body to the `.container`
- `index.njk` contains the content found inside the `.container`

A `login.njk` page has also been introduced that provides a link to the
GitHub login webhook.