Implement logging in using GitHub OAuth
User tokens are stored as a cookie. If no valid token is presented, the
user is shown the login page, which allows them to authorise Towtruck to
access their account. This then redirects to the Towtruck callback URL,
which generates the token and instructs the browser to store it before
redirecting to the index page.