Allow deleting of all data with a particular scope from the database
Currently there is no method to delete all data from the database with a
particular scope (e.g. all repositories from all organisations). This is
desirable to have to ensure that the database is in a known good state
when seeding, especially when something like changing how key columns
are interpreted occurs.