Allow Playwright tests to be authenticated
The `GH_TOKEN` environment variable can be used both locally and in CI
to avoid testing the GitHub OAuth flow (which is provided by the Octokit
library and can therefore be assumed to be well tested).

In local development the token can be retrieved by logging into Towtruck
and then copying the value of the `Token` cookie.

In CI, the correct value is automatically populated by GitHub Actions
into a secret.