Implement username and password login
This will mainly be used for end-to-end testing in the CI pipeline to
avoid having to authenticate with GitHub, although this implementation
corresponds to best practices, in that passwords are not stored in
plaintext, a salt is used, and hashes are computed using a secure
algorithm (SHA512).

Currently new users are not created directly within the application,
although a unique random salt should be provided for each user when this
is implemented.