Implement logging out the current user
This involves creating a custom middleware to set the current user on
the request, which runs after the JWT auth is verified, allowing the
application to retrieve the name of the current user.

If the current user is set, the templating engine will display a welcome
message to the user and also provide a link to log out. If clicked, the
server instructs the browser to immediately expire the token cookie,
with the effect of logging them out.

To make sure that this works with GitHub OAuth, the GitHub login
callback now stores a `User` object representing the GitHub user in the
database, using the organisations that they're subscribed to at the time
of login.