Implement index page
The root path `/` now displays a list of organisations using Towtruck
that the user is a member of. Items in the list link to
organisation-specific dashboards available at `/orgs/:org`. Users that
aren't members of an organisation will see a 404 page, preventing the
unauthorised access to information about private repositories.