Expand Playwright tests
Several new end-to-end tests have been introduced to expand the
confidence we have in the overall functioning of Towtruck, especially
considering that the application has grown significantly more complex.

Tests are now roughly scoped in their files according to the aspect of
the application being tested (e.g. logging in).