Support nodenv
It's a bit more common for us to use nodenv at dxw than nvm, and it can
be a pain having both nvm and nodenv managing Node installations on the
same system

This therefore adds support for using nodenv locally without dropping
nvm support. We check whether nvm or nodenv are installed, and use the
appropriate one to install Node. If neither are installed, we error and
exit the script