Check --seed argument in script/bootstrap
Currently any argument will lead `script/bootstrap` to run the seed
process. Per the documentation, we should check that the argument is
`--seed`

We are only checking the first argument (not counting `script/bootstrap`
as an argument). If we later want to support more arguments/arguments in
arbitrary positions, we could use a format like one of these:
- https://github.com/dxw/rails-template/blob/d2719a6/script/all/process-script-args#L9-L25
- https://github.com/yndajas/dotfiles/blob/0e35eeb/src/zsh/.zshrc#L163-L169