Create `data` folder on setup
This is needed for the seed script to run successfully. The default
filepath in the `TowtruckDatabase` contructor in `db/script.js` is in
this folder, but the folder is empty by default (there are no committed
files in there). In `initialiseDatabase` an error is raised when trying
to create the database in a non-existent folder