Make links relative in sortable table header Nunjucks macro
The sortableTableHeader macro will generate links that use query
parameters to control the sorting of the dashboard. However, as the
table header links are of the form
/?sortBy=<sort>&sortDirection=<direction>, they will only work
correctly when used on the root path. To make them usable anywhere,
just the query part of the URL can be used as the href, which will
link to the current page with the updated query parameters.