Refactor Nunjucks to be used as the Express rendering engine
Express introduces the concepts of views and rendering, which simplifies
the process of producing an HTML response. It's straightforward to use
Nunjucks as the rendering engine as the `nunjucks.render` method has the
same signature as the `callback` passed to `Express.engine`.

The `views/` folder is now the location in which Nunjucks templates
should be stored, and it currently contains:

- `_macros.njk`: basic repeatable utilities
- `index.njk`: the repository dashboard