Introduce a 404 page
When a user tries to navigate to a page that doesn't exist, a 404
response should be presented. Most server frameworks handle this
automatically, but the default page presented won't look like part of
the application.

To make the 404 page look like the rest of the application, a new view
has been introduced that uses the base template. A catch-all route has
been defined that will render this view with the path that the user
attempted to navigate to, an explanatory message, and a link back to the
home page.